Correction of  downloadAsPDF method. Fixed download or print issue in Android.
In the functionality of download button I used printWindow.print(); which works fine in Windows, Mac, and IOS, but not in Android. So, to solve this I first external libraries like html2pdf which includes html2canvas and jsPDF. It almost worke d as it was generating blank output. Searched for solutions but still not desired output. Finally  placed the content in iframe  and  increased setTimeout for better UX. This worked successfully in both android and windows.